Renamed features folder to lib
